move integer tests to own file
